feat(xtask): cargo check-style 现在和 CI/build 一致
