+++ xtask/README.md
@@ -34,19 +34,3 @@
   - [x] 打包可烧写到其他存储介质的 zCore 镜像
   - [ ] 启动与 Qemu 关联的 GDB 以支持内核调试
 - [ ] 自动化测试：实现与 CI 的逻辑一致性
-
-## 进度日志
-
-- 2022/05/06
-
-  顶层的 Makefile 已经尽量迁移到 rust，并在子项目 README.md 中更新了子命令说明。
-
-  计划提起一次 PR。
-
-- 2022/05/04
-
-  初步的计划是先尽量将 Makefile 转化为类型安全且更有可能工程化结构化的 Rust xtask。
-  尤其是要将 zCore 目录内外的两个 Makefile 合并。
-
-  目前已经架空了外面 Makefile 的 rootfs 指令，这个指令是用于将加载到内存的最小系统的。
-  外面的 Makefile 还剩打包镜像、启动某些测试集的功能，但目前命令之间不正交，还需要进一步梳理。
